+++ lib/src/sanitizable.dart
@@ -1,0 +1,4 @@
+/// Sanitizable interface
+abstract class Sanitizable {
+  void sanitize();
+}