+++ lib/src/types.dart
@@ -1,0 +1,2 @@
+typedef Validator = String? Function(String?);
+typedef Sanitizer = String? Function(String?);