+++ lib/src/validation_chain_api.dart
@@ -1,14 +1,14 @@
-typedef Validator = String? Function(String?);
+import 'types.dart';
 
-/// Validation Chain API
-///
-/// The validation chain api to use with [TextFormField] in Flutter or Backend applicaitons made with Dart.
-///
-/// You can add as many validators to a chain as you need.
-/// When the `validate` function is called, it will run each [Validator] in the order they were specified.
 class ValidationChain {
   final List<Validator> validators;
 
+  /// Validation Chain API
+  ///
+  /// The validation chain api to use with [TextFormField] in Flutter or Backend applicaitons made with Dart.
+  ///
+  /// You can add as many validators to a chain as you need.
+  /// When the `validate` function is called, it will run each [Validator] in the order they were specified.
   const ValidationChain(this.validators);
 
   /// runs each [Validator] in the order they were specified.