+++ lib/validation_chain.dart
@@ -1,3 +1,7 @@
 library validation_chain;
 
+export 'src/map_sanitizer.dart';
+export 'src/map_validator.dart';
+export 'src/sanitizable.dart';
+export 'src/types.dart';
 export 'src/validation_chain_api.dart';